AI Research Assistant/requirements.txt
# AI Research Assistant - Python Dependencies
# Install all with:  pip install -r requirements.txt

# ── Core AI Framework ──────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── LLM Providers ─────────────────────────────────────────────────────────────
langchain-groq>=0.2.0          # Groq LLM (fast & free)
langchain-google-genai>=2.0.0  # Google Gemini (optional alternative)

# ── Web Search ────────────────────────────────────────────────────────────────
tavily-python>=0.3.0           # Tavily web search tool

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic>=2.0.0
typing_extensions>=4.0.0

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.40.0

# ── Environment Variables ─────────────────────────────────────────────────────
python-dotenv>=1.0.0
